⚡ :: AccountInfoViewModel 작성
